requirements.txt
absl-py==0.11.0
aiohttp==3.7.2
aiohttp-cors==0.7.0
aioredis==1.3.1
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
avalanche-lib==0.1.0
beautifulsoup4==4.9.3
blessings==1.7
blis==0.7.4
cached-property==1.5.2
cachetools==4.1.1
catalogue==2.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cmake==3.21.0
colorama==0.4.4
colorful==0.5.4
contextvars==2.4
cycler==0.10.0
cymem==2.0.5
Cython==0.29.28
dataclasses==0.6
distro==1.7.0
docker-pycreds==0.4.0
fastcore==1.3.19
fastprogress==1.0.0
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gdown==4.4.0
gitdb==4.0.9
GitPython==3.1.27
google==3.0.0
google-api-core==1.23.0
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
gpustat==0.6.0
GPUtil==1.4.0
grpcio==1.32.0
h5py==2.10.0
hiredis==1.1.0
idna==2.10
idna-ssl==1.1.0
imbalanced-learn==0.8.0
immutables==0.14
importlib-metadata==1.7.0
install==1.3.4
Jinja2==2.11.3
joblib==0.17.0
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
liac-arff==2.5.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
msgpack==1.0.0
multidict==5.0.0
murmurhash==1.0.5
nptyping==1.4.4
numpy>=1.15.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencensus==0.7.11
opencensus-context==0.1.2
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.4
pathtools==0.1.2
pathy==0.4.0
Pillow==7.2.0
preshed==3.0.5
prometheus-client==0.8.0
promise==2.3
protobuf==3.12.4
psutil==5.7.3
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pydantic==1.7.3
pykeops==1.5
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-ranger==0.1.1
pytorchcv==0.0.67
pytz==2020.4
PyYAML==5.3.1
quadprog==0.1.11
ray==1.0.1
redis==3.4.1
requests==2.24.0
requests-oauthlib==1.3.0
river==0.7.0
rsa==4.6
scikit-build==0.13.1
scikit-learn==0.24.1
scikit-multiflow==0.5.3
scipy==1.5.4
sentry-sdk==1.5.8
setproctitle==1.2.2
shortuuid==1.0.8
six==1.15.0
smart-open==3.0.0
smmap==5.0.0
sortedcontainers==2.3.0
soupsieve==2.0.1
spacy==3.0.3
spacy-legacy==3.0.1
srsly==2.4.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
thinc==8.0.1
threadpoolctl==2.1.0
torch==1.10.1
torch-optimizer==0.1.0
torchsummary==1.5.1
torchvision==0.11.2
tqdm==4.52.0
typer==0.3.2
typing-extensions==3.7.4.3
typish==1.9.3
urllib3==1.25.10
wandb==0.12.11
wasabi==0.8.2
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
yarl==1.6.2
yaspin==2.1.0
zipp==3.1.0
